Send data continuously and implement refresh timer
Set up a timer for rx and tx so receive and send don't sleep.
Refresh happens every one second to ensure code is still running
Error shows that in rx, it stops receiving after a bit and stops
updating display. The IP update in tx was also moved as it was
incorrectly placed in the code