Send data and update screen once a second
This code was updated to try addressing the code crashing issue
in the tx_temp_data.py. It was revealed that a recursionerror was
to blame, that the function was reaching recursion limit and the
functions called in an endless loop were those from the adafruit
package relating to populating the display. Therefore, it seems
that overupdating the display may be to blame.

As in practice data will only be sent intermittenly, it seemed
fair to have the display update intermittenly as well. Therefore,
this was the proposed solution to the random display crashing issue