Update data written to csv file
Chiefly, the epoch timestamps written in the csv file are convert-
ed to human-readable time stamps. This includes the time received
as well as the timestamp in the received packet and the log time
start.
Additionally, data is written for the seconds when no packets are
received. This will track wether the node at the BDR station dies
first, or if the node at the lab dies first as right now its un-
clear.